requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.11
python-dotenv==1.0.0